refactor the code based on review
