+++ tools/automatic_query_fixer/src/main/java/com/google/cloud/bigquery/utils/queryfixer/errors/BigQuerySemanticError.java
@@ -3,12 +3,10 @@ package com.google.cloud.bigquery.utils.queryfixer.errors;
 import com.google.cloud.bigquery.BigQueryException;
 import com.google.cloud.bigquery.utils.queryfixer.entity.Position;
 
-
-/** A class to represent the semantic errors of BigQuery.*/
+/** A class to represent the semantic errors of BigQuery. */
 public abstract class BigQuerySemanticError extends BigQuerySqlError {
 
   public BigQuerySemanticError(Position errPos, BigQueryException errorSource) {
     super(errPos, errorSource);
   }
-
 }
