+++ tools/automatic_query_fixer/src/main/java/com/google/cloud/bigquery/utils/queryfixer/errors/BigQuerySqlError.java
@@ -5,17 +5,22 @@ import com.google.cloud.bigquery.utils.queryfixer.entity.Position;
 
 import lombok.Getter;
 
-/** The base class to represent the SQL Error generated by the BigQuery server. It should be resolved from
- * {@link BigQueryException}. */
+/**
+ * The base class to represent the SQL Error generated by the BigQuery server. It should be resolved
+ * from {@link BigQueryException}.
+ */
 @Getter
 public abstract class BigQuerySqlError {
 
-  BigQueryException errorSource;
-  /** The occurrence of the error at a query. It may be null if the error does not belong to any specific locations. */
-  Position errPos;
+  private final BigQueryException errorSource;
+  /**
+   * The occurrence of the error at a query. It may be null if the error does not belong to any
+   * specific locations.
+   */
+  private final Position errorPosition;
 
-  public BigQuerySqlError(Position errPos, BigQueryException errorSource) {
-    this.errPos = errPos;
+  public BigQuerySqlError(Position errorPosition, BigQueryException errorSource) {
+    this.errorPosition = errorPosition;
     this.errorSource = errorSource;
   }
 }
