+++ tools/automatic_query_fixer/src/main/java/com/google/cloud/bigquery/utils/queryfixer/errors/FunctionNotFoundError.java
@@ -5,27 +5,34 @@ import com.google.cloud.bigquery.utils.queryfixer.entity.Position;
 
 import lombok.Getter;
 
-/** A class to represent the "Function Not Found" errors. The errors look like one of the two forms: (1) Function not
- * found: [function] at [x:y] or (2) Function not found: [function]; Did you mean [suggestion]? at [x:y], where
- * [function] is the incorrect function, [suggestion] is one of the correct function candidates, and [x:y] is the row
- * and column numbers. */
+/**
+ * A class to represent the "Function Not Found" errors. The errors look like one of the two forms:
+ *
+ * <ol>
+ *   <li>(1) Function not found: [function] at [x:y]
+ *   <li>(2) Function not found: [function]; Did you mean [suggestion]? at [x:y]
+ * </ol>
+ *
+ * <p>where [function] is the incorrect function, [suggestion] is one of the correct function
+ * candidates, and [x:y] is the row and column numbers.
+ */
 @Getter
 public class FunctionNotFoundError extends BigQuerySemanticError {
 
-  private final String incorrectFunctionName;
+  private final String functionName;
   private final String suggestion;
 
-
-  public FunctionNotFoundError(String incorrectFunctionName, Position errPos, String suggestion,
-      BigQueryException errorSource) {
+  public FunctionNotFoundError(
+      String functionName, Position errPos, String suggestion, BigQueryException errorSource) {
     super(errPos, errorSource);
-    this.incorrectFunctionName = incorrectFunctionName;
+    this.functionName = functionName;
     this.suggestion = suggestion;
   }
 
-  public FunctionNotFoundError(String incorrectFunctionName, Position errPos, BigQueryException errorSource) {
+  public FunctionNotFoundError(
+      String functionName, Position errPos, BigQueryException errorSource) {
     super(errPos, errorSource);
-    this.incorrectFunctionName = incorrectFunctionName;
+    this.functionName = functionName;
     this.suggestion = null;
   }
