+++ tools/automatic_query_fixer/src/main/java/com/google/cloud/bigquery/utils/queryfixer/errors/TableNotFoundError.java
@@ -5,16 +5,18 @@ import com.google.cloud.bigquery.utils.queryfixer.entity.Position;
 
 import lombok.Getter;
 
-/** A class to represent the "Table Not Found" errors from BigQuery. The errors are presented in this form:
- * "Not found: Table [TableName] was not found", where TableName is the incorrect table name. */
+/**
+ * A class to represent the "Table Not Found" errors from BigQuery. The errors are presented in this
+ * form: "Not found: Table [TableName] was not found", where [TableName] is the incorrect table
+ * name.
+ */
 @Getter
 public class TableNotFoundError extends BigQuerySemanticError {
 
-  String incorrectTableName;
+  private final String tableName;
 
-  public TableNotFoundError(String incorrectTableName, Position errPos, BigQueryException errorSource) {
+  public TableNotFoundError(String tableName, Position errPos, BigQueryException errorSource) {
     super(errPos, errorSource);
-    this.incorrectTableName = incorrectTableName;
+    this.tableName = tableName;
   }
-
 }
